+++ app/src/main/java/com/example/quicknotes/UI/notes/EditNoteDialog.kt
@@ -1,5 +1,9 @@
 package com.example.quicknotes.UI.notes
 
+import androidx.compose.animation.animateColor
+import androidx.compose.animation.core.animateDp
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.core.updateTransition
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
@@ -17,8 +21,8 @@ import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Close
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material.icons.filled.Clear
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
@@ -29,6 +33,7 @@ import androidx.compose.material3.OutlinedTextFieldDefaults
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -36,17 +41,16 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Outline
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import androidx.compose.ui.window.DialogProperties
 import com.example.quicknotes.data.notesData.NotesEntity
 import com.example.quicknotes.viewmodel.NotesViewModel
+import kotlinx.coroutines.delay
 import java.text.SimpleDateFormat
 import java.util.Date
 import java.util.Locale
-
 
 @Composable
 fun EditNoteDialog(
@@ -61,6 +65,30 @@ fun EditNoteDialog(
         unfocusedBorderColor = Color.Transparent,
         cursorColor = MaterialTheme.colorScheme.primary
     )
+    var deleteButton by remember { mutableStateOf(false) }
+    var clearButton by remember { mutableStateOf(false) }
+    val transition = updateTransition(targetState = deleteButton, label = "deleteButtonTransition")
+    val transition2 = updateTransition(targetState = clearButton, label = "clearButtonTransition")
+
+    val buttonColor by transition.animateColor(
+        label = "buttonColor",
+        transitionSpec = { tween(durationMillis = 100) }
+    ) { isDelete ->
+        if (isDelete) Color.Red else Color.LightGray
+    }
+
+    val deleteButtonWidth by transition.animateDp(
+        label = "buttonWidth",
+        transitionSpec = { tween(durationMillis = 300) }
+    ) { isDelete ->
+        if (isDelete) 130.dp else 70.dp
+    }
+    val clearButtonWidth by transition2.animateDp(
+        label = "buttonWidth",
+        transitionSpec = { tween(durationMillis = 300) }
+    ) { isClear ->
+        if (isClear) 130.dp else 70.dp
+    }
     Dialog(
         onDismissRequest = onDismiss,
         properties = DialogProperties(usePlatformDefaultWidth = false)
@@ -85,46 +113,72 @@ fun EditNoteDialog(
                             .clickable(
                                 onClick = onDismiss
                             ),
-                        imageVector = Icons.Default.Close,
+                        imageVector = Icons.Default.ArrowBack,
                         contentDescription = null,
                     )
                     Row(Modifier.wrapContentSize()){
                         Button(
                             modifier = Modifier
-                                .wrapContentSize(),
+                                .width(deleteButtonWidth).height(40.dp),
                             onClick = {
-                                notesViewModel.delete(notesEntity)
-                                onDismiss()
+                                if (deleteButton){
+                                    notesViewModel.delete(notesEntity)
+                                    onDismiss()
+                                }else {
+                                    deleteButton = true
+                                }
                             },
-                            colors = ButtonDefaults.buttonColors(containerColor = Color.LightGray),
+                            colors = ButtonDefaults.buttonColors(containerColor = buttonColor),
                             shape = RoundedCornerShape(20.dp, 5.dp, 5.dp, 20.dp)
                         )
                         {
-                            Image(
+                            Icon(
                                 imageVector = Icons.Default.Delete,
-                                contentDescription = "Delete"
+                                contentDescription = "Delete",
+                                modifier = Modifier
+                                    .size(24.dp)
                             )
+                            if (deleteButton) {
+                                Text(
+                                    text = "Delete?",
+                                    modifier = Modifier.padding(horizontal = 8.dp)
+                                )
+                            }
+                        }
+                        LaunchedEffect(deleteButton || clearButton) {
+                            if (deleteButton || clearButton) {
+                                delay(2500)
+                                deleteButton = false
+                                clearButton = false
+                            }
                         }
                         Spacer(Modifier.width(5.dp))
                         Button(
-                            modifier = Modifier
-                                .wrapContentSize(),
+                            modifier = Modifier.width(clearButtonWidth)
+                                .height(40.dp),
                             onClick = {
-                                val updateNote = notesEntity.copy(
-                                    title = title,
-                                    description = description
-                                )
-                                notesViewModel.update(updateNote)
-                                onDismiss()
+                                if (clearButton){
+                                    title = ""
+                                    description = ""
+                                    clearButton = false
+                                }else{
+                                    clearButton = true
+                                }
                             },
                             colors = ButtonDefaults.buttonColors(containerColor = Color.LightGray),
                             shape = RoundedCornerShape(5.dp, 20.dp, 20.dp, 5.dp)
                         )
                         {
                             Image(
-                                imageVector = Icons.Default.Check,
-                                contentDescription = "Save"
+                                imageVector = Icons.Default.Clear,
+                                contentDescription = "Clear"
                             )
+                            if (clearButton){
+                                Text(
+                                    text = "Clear?",
+                                    modifier = Modifier.padding(horizontal = 8.dp)
+                                )
+                            }
                         }
                     }
 
@@ -168,4 +222,9 @@ fun EditNoteDialog(
             }
         }
     }
+    val updateNote = notesEntity.copy(
+        title = title,
+        description = description
+    )
+    notesViewModel.update(updateNote)
 }
